Pramod api

+++ app/components/Header/UserMenu.js
@@ -1,44 +1,45 @@
-import React, { useState } from "react";
-import PropTypes from "prop-types";
-import { withStyles, makeStyles } from "@material-ui/core/styles";
-import classNames from "classnames";
-import { Link } from "react-router-dom";
-import Avatar from "@material-ui/core/Avatar";
-import Tooltip from "@material-ui/core/Tooltip";
-import IconButton from "@material-ui/core/IconButton";
-import Button from "@material-ui/core/Button";
-import Info from "@material-ui/icons/Info";
-import Warning from "@material-ui/icons/Warning";
-import Check from "@material-ui/icons/CheckCircle";
-import Error from "@material-ui/icons/RemoveCircle";
-import ExitToApp from "@material-ui/icons/ExitToApp";
-import Badge from "@material-ui/core/Badge";
-import Divider from "@material-ui/core/Divider";
-import Menu from "@material-ui/core/Menu";
-import MenuItem from "@material-ui/core/MenuItem";
-import ListItemIcon from "@material-ui/core/ListItemIcon";
-import ListItemText from "@material-ui/core/ListItemText";
-import ListItemAvatar from "@material-ui/core/ListItemAvatar";
-import Drawer from "@material-ui/core/Drawer";
-import List from "@material-ui/core/List";
-import ListItem from "@material-ui/core/ListItem";
-import ListItemSecondaryAction from "@material-ui/core/ListItemSecondaryAction";
-import dummy from "dan-api/dummy/dummyContents";
-import messageStyles from "dan-styles/Messages.scss";
-import avatarApi from "dan-api/images/avatars";
-import companyLogo from "dan-api/images/logos";
-import link from "dan-api/ui/link";
-import styles from "./header-jss";
+import React, { useState } from 'react';
+import PropTypes from 'prop-types';
+import { withStyles, makeStyles } from '@material-ui/core/styles';
+import classNames from 'classnames';
+import { Link } from 'react-router-dom';
+import Avatar from '@material-ui/core/Avatar';
+import Tooltip from '@material-ui/core/Tooltip';
+import IconButton from '@material-ui/core/IconButton';
+import Button from '@material-ui/core/Button';
+import Info from '@material-ui/icons/Info';
+import Warning from '@material-ui/icons/Warning';
+import Check from '@material-ui/icons/CheckCircle';
+import Error from '@material-ui/icons/RemoveCircle';
+import ExitToApp from '@material-ui/icons/ExitToApp';
+import Badge from '@material-ui/core/Badge';
+import Divider from '@material-ui/core/Divider';
+import Menu from '@material-ui/core/Menu';
+import MenuItem from '@material-ui/core/MenuItem';
+import ListItemIcon from '@material-ui/core/ListItemIcon';
+import ListItemText from '@material-ui/core/ListItemText';
+import ListItemAvatar from '@material-ui/core/ListItemAvatar';
+import Drawer from '@material-ui/core/Drawer';
+import List from '@material-ui/core/List';
+import ListItem from '@material-ui/core/ListItem';
+import ListItemSecondaryAction from '@material-ui/core/ListItemSecondaryAction';
+import dummy from 'dan-api/dummy/dummyContents';
+import messageStyles from 'dan-styles/Messages.scss';
+import avatarApi from 'dan-api/images/avatars';
+import companyLogo from 'dan-api/images/logos';
+import link from 'dan-api/ui/link';
+import styles from './header-jss';
 
-import "../../styles/custom/customheader.css";
-import { LOGIN_URL } from "../../utils/constants";
+import '../../styles/custom/customheader.css';
+import { access_token, ACCOUNT_API_URL, LOGIN_URL } from '../../utils/constants';
+import axios from 'axios';
 
 const useStyles = makeStyles({
   list: {
     width: 300,
   },
   fullList: {
-    width: "auto",
+    width: 'auto',
   },
 });
 
@@ -77,8 +78,27 @@ function UserMenu(props) {
   };
 
   const handleLogout = () => {
-    localStorage.clear();
-    window.location.href = LOGIN_URL;
+    console.log(access_token);
+    const config = {
+      method: 'get',
+      url: `${ACCOUNT_API_URL}api/v1/user/logout/`,
+      headers: {
+        Authorization: `Bearer ${access_token}`,
+      }
+    };
+    console.log(config);
+    axios(config)
+      .then((response) => {
+        if (response.status === 201) {
+          localStorage.clear();
+          window.location.href = 'https://dev-accounts-api.paceos.io/login';
+        }
+      })
+      .catch((error) => {
+        console.log(error);
+      });
+
+    // window.location.href = LOGIN_URL;
   };
 
   function ListItemLink(props) {
@@ -102,7 +122,7 @@ function UserMenu(props) {
 
       <IconButton
         aria-haspopup="true"
-        onClick={handleMenu("notification")}
+        onClick={handleMenu('notification')}
         color="inherit"
         className={classNames(
           classes.notifIcon,
@@ -117,12 +137,12 @@ function UserMenu(props) {
         id="menu-notification"
         anchorEl={anchorEl}
         anchorOrigin={{
-          vertical: "top",
-          horizontal: "right",
+          vertical: 'top',
+          horizontal: 'right',
         }}
         transformOrigin={{
-          vertical: "bottom",
-          horizontal: "left",
+          vertical: 'bottom',
+          horizontal: 'left',
         }}
         className={classes.notifMenu}
         PaperProps={{
@@ -130,7 +150,7 @@ function UserMenu(props) {
             width: 350,
           },
         }}
-        open={openMenu === "notification"}
+        open={openMenu === 'notification'}
         onClose={handleClose}
       >
         <MenuItem onClick={handleClose}>
@@ -313,7 +333,7 @@ function UserMenu(props) {
 
       <Button
         className={classes.userControls}
-        onClick={handleMenu("user-setting")}
+        onClick={handleMenu('user-setting')}
       >
         <img className={classes.userLogo} src={companyLogo[0]} />
         <Avatar
@@ -326,14 +346,14 @@ function UserMenu(props) {
         id="menu-appbar"
         anchorEl={anchorEl}
         anchorOrigin={{
-          vertical: "top",
-          horizontal: "right",
+          vertical: 'top',
+          horizontal: 'right',
         }}
         transformOrigin={{
-          vertical: "top",
-          horizontal: "right",
+          vertical: 'top',
+          horizontal: 'right',
         }}
-        open={openMenu === "user-setting"}
+        open={openMenu === 'user-setting'}
         onClose={handleClose}
       >
         <MenuItem onClick={handleClose} component={Link} to={link.profile}>
